refactor Blog code and its ten server
